Add files via upload
This is a basic code for a RAG Chatbot. Need help trouble shooting. I used Jupyter Notebook. The problem I am having is in the output. If you ask a question, you will see what I mean. I have uploaded word.docx which is where I have cleaned up transcripts of some of the more recent Trump rallies. This is what the Code is pulling from, so it needs to be in the main directory or the file path will have to be updated accordingly.